+++ overrides/app/Http/Requests/Base/LocaleRequest.php
@@ -10,7 +10,7 @@ class LocaleRequest extends FormRequest
     public function rules(): array
     {
         return [
-            'locale' => ['required', 'string', Rule::in(['en', 'de', 'swg', 'bar'])],
+            'locale' => ['required', 'string', Rule::in(__PTERO_I18N_LOCALE_ALLOWLIST__)],
             'namespace' => ['required', 'string', 'regex:/^[a-z]{1,191}$/'],
         ];
     }